% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_clustering.R
\name{clustering_iteration}
\alias{clustering_iteration}
\title{Performs an iteration of Leiden clustering to a Seurat object.}
\usage{
clustering_iteration(seurat_object, min_score, min_size)
}
\arguments{
\item{seurat_object}{an SCTransformed, integrated Seurat object}

\item{min_score}{minimum clustering score}

\item{min_size}{minimum cluster size}
}
\value{
a Seurat object with clusters in the "leiden_clusters" slot
}
\description{
Performs an iteration of Leiden clustering to a Seurat object.
}
