% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_clustering.R
\name{iterative_clustering}
\alias{iterative_clustering}
\title{Performs iterative clustering on a Seurat object to find clusters expressing significant DE genes.}
\usage{
iterative_clustering(
  seurat_object,
  max_iterations = 10,
  min_score = 150,
  cluster_size = 20
)
}
\arguments{
\item{seurat_object}{an SCTransformed, integrated Seurat object}

\item{max_iterations}{number of clustering iterations to perform}

\item{min_score}{minimum cluster score}

\item{cluster_size}{minimum cluster size}
}
\value{
a Seurat object with clusters in the "leiden_clusters" slot
}
\description{
Performs iterative clustering on a Seurat object to find clusters expressing significant DE genes.
}
