% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_clustering.R
\name{leiden_clustering}
\alias{leiden_clustering}
\title{Leiden clustering into a set number of clusters. Varies resolution until the number of clusters is achieved.
Requires a conda environment with igraph and leidenalg installed.}
\usage{
leiden_clustering(
  seurat_object,
  num_clusters = 2,
  score_limit = 150,
  min_size = 20
)
}
\arguments{
\item{seurat_object}{an SCTransformed, integrated Seurat object}

\item{num_clusters}{number of desired clusters. Default = 2.}

\item{score_limit}{minimum score for a cluster. Default = 150.}

\item{min_size}{minimum size of a cluster. Default = 20.}
}
\value{
a Seurat object with clusters in the "leiden_clusters" slot
}
\description{
Leiden clustering into a set number of clusters. Varies resolution until the number of clusters is achieved.
Requires a conda environment with igraph and leidenalg installed.
}
