% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_clustering.R
\name{clustering_iteration}
\alias{clustering_iteration}
\title{Performs an iteration of Leiden clustering to a Seurat object.}
\usage{
clustering_iteration(seurat_object, min_score, cluster_size, pct.1)
}
\arguments{
\item{seurat_object}{a normalized, integrated Seurat object}

\item{min_score}{minimum clustering score}

\item{cluster_size}{minimum cluster size}

\item{pct.1}{fraction of gene expression in the overexpressing cluster}
}
\value{
a Seurat object with clusters in the "leiden_clusters" slot
}
\description{
Performs an iteration of Leiden clustering to a Seurat object.
}
